chore: move controlplanestack inside stacks dir
